#10 Enroll a student abroad in a course
The admin in the destiny administrative office can check the requests in the link Scheduling -> Universities Requests Abroad. If the student is accepted, the admin user needs to choose the period of the course.
S/he also needs to select the attendance start date and, after clicking the "Enroll Students Abroad" button, the system will enroll the student in the destiny administrative system and the learning environment system.